refactor(#53): 외부 서비스 패키지 분리
